+++ src/cli.js
@@ -3,7 +3,7 @@ const commandLineUsage = require('command-line-usage');
 const Vault = require('./vault').Vault;
 
 const init = ({ path }) => {
-  const vault = new Vault({ credentialsFilePath: `${path}/credentials.json` });
+  const vault = new Vault({ credentialsFilePath: path });
   if (fs.existsSync(`${vault.credentialsFilePath}.enc`)) {
     console.log('Warning credentials.json.enc exists, ensure decrypt file before generate new key');
   } else if (fs.existsSync(`${vault.credentialsFilePath}.key`)) {
@@ -15,30 +15,29 @@ const init = ({ path }) => {
       .then(() => {
         fs.unlinkSync(`${vault.credentialsFilePath}`);
       })
-      .catch((error) => {});
+      .catch((error) => console.error(error));
   }
 };
 
 const encrypt = ({ path }) => {
-  const vault = new Vault({
-    credentialsFilePath: `${path}/credentials.json`,
-  });
+  const vault = new Vault({ credentialsFilePath: path });
   if (fs.existsSync(`${vault.credentialsFilePath}.key`) || process.env.NODE_MASTER_KEY) {
     vault
       .encryptFile()
       .then(() => {
         fs.unlinkSync(`${vault.credentialsFilePath}`);
+        try {
+          fs.unlinkSync(`${vault.credentialsFilePath}.iv`);
+        } catch {}
       })
-      .catch((error) => {});
+      .catch((error) => console.error(error));
   } else {
     console.log('Warning credentials.json.key not exists, create new key with init');
   }
 };
 
 const edit = ({ path }) => {
-  const vault = new Vault({
-    credentialsFilePath: `${path}/credentials.json`,
-  });
+  const vault = new Vault({ credentialsFilePath: path });
   if (!fs.existsSync(`${vault.credentialsFilePath}.enc`)) {
     console.log('Error credentials.json.enc not exists');
   } else {
@@ -82,5 +81,6 @@ module.exports = {
   init,
   encrypt,
   edit,
+  decrypt: edit,
   help,
 };