+++ src/vault.js
@@ -1,25 +1,43 @@
 const core = require('./core');
 const render = require('./template-render').render;
-var fs = require('fs');
+const fs = require('fs');
+const { get } = require('lodash');
 
 class Vault {
   constructor({
-    decryptFnc = core.decrypt,
-    encryptFnc = core.encrypt,
+    decryptFnc = core.decryptJSON,
+    encryptFnc = core.encryptJSON,
     credentialsFilePath = 'credentials.json',
     nodeEnv = 'development',
+    masterKey,
   } = {}) {
     this.decryptFnc = decryptFnc;
     this.encryptFnc = encryptFnc;
     this.credentialsFilePath = credentialsFilePath;
-    this.credentials = {};
-    this.credentialsEnv = {};
+    this.masterKey = masterKey;
+    this._credentials = {};
+    this._credentialsEnv = {};
     this.nodeEnv = nodeEnv === '' ? 'development' : nodeEnv;
+    this.configured = false;
+  }
+
+  get credentials() {
+    if (!this.configured) {
+      this.config();
+    }
+    return this._credentials;
+  }
+
+  get credentialsEnv() {
+    if (!this.configured) {
+      this.config();
+    }
+    return this._credentialsEnv;
   }
 
   setCredentials(credentials) {
-    this.credentials = { ...credentials };
-    this.credentialsEnv = { ...credentials[this.nodeEnv] };
+    this._credentials = { ...credentials };
+    this._credentialsEnv = get(credentials, this.nodeEnv, {});
   }
 
   config({ keyValue, path } = {}) {
@@ -28,17 +46,24 @@ class Vault {
     }
     const key = keyValue || this.getMasterKey();
     const text = fs.readFileSync(`${this.credentialsFilePath}.enc`, 'utf8');
-    const credentialsText = this.decryptFnc(key, text);
+    const [credentialsText, iv] = this.decryptFnc(key, text);
     const credentialsTextRendered = render(credentialsText);
     const credentials = JSON.parse(credentialsTextRendered);
     this.setCredentials(credentials);
+    this.configured = true;
     return credentials;
   }
 
   async encryptFile({ keyValue } = {}) {
     const key = keyValue || this.getMasterKey();
     const text = fs.readFileSync(this.credentialsFilePath, 'utf8').trim();
-    const cipherBundle = await this.encryptFnc(key, text);
+    let iv;
+    try {
+      iv = fs.readFileSync(`${this.credentialsFilePath}.iv`, 'utf8').trim();
+    } catch {
+      iv = null;
+    }
+    const cipherBundle = await this.encryptFnc(key, text, iv);
     fs.writeFileSync(`${this.credentialsFilePath}.enc`, cipherBundle);
     return `${this.credentialsFilePath}.enc`;
   }
@@ -46,8 +71,9 @@ class Vault {
   editCredentials({ keyValue } = {}) {
     const key = keyValue || this.getMasterKey();
     const text = fs.readFileSync(`${this.credentialsFilePath}.enc`, 'utf8');
-    const decryptCredentials = this.decryptFnc(key, text);
+    const [decryptCredentials, iv] = this.decryptFnc(key, text);
     fs.writeFileSync(`${this.credentialsFilePath}`, decryptCredentials, 'utf8');
+    fs.writeFileSync(`${this.credentialsFilePath}.iv`, iv, 'utf8');
     return this.credentialsFilePath;
   }
 
@@ -58,7 +84,9 @@ class Vault {
   }
 
   getMasterKey() {
-    return process.env.NODE_MASTER_KEY || fs.readFileSync(`${this.credentialsFilePath}.key`, 'utf8').trim();
+    return (
+      this.masterKey || process.env.NODE_MASTER_KEY || fs.readFileSync(`${this.credentialsFilePath}.key`, 'utf8').trim()
+    );
   }
 }
 module.exports.Vault = Vault;