+++ test/core.spec.js
@@ -1,5 +1,6 @@
 const core = require('../src/core');
 require('./helpers/matchers');
+const complexJSON = require('./examples/transformValues/complex.json');
 
 describe('core', () => {
   let NODE_MASTER_KEY = '8aa93853b3ff01c5b5447529a9c33cb9';
@@ -10,11 +11,26 @@ describe('core', () => {
     expect(result).validEncrypted();
   });
 
-  test('decrypt', () => {
-    const result = core.decrypt(
+  test('encrypt with any key length', async () => {
+    const result = await core.encrypt('12345678', credentialsString);
+    expect(result).validEncrypted();
+  });
+
+  test('decrypt', async () => {
+    const [result, iv] = await core.decrypt(
       NODE_MASTER_KEY,
       'dwAhexc3PhUrX9i4gutpy6Hb8endKm7hMCQALPspYEc=--84X822lxzoPbO9Jh2knEGA=='
     );
     expect(JSON.parse(result)).toEqual(credentials);
+  });
+
+  test('transformValues sync', () => {
+    const result = core.transformValues(complexJSON, (value) => `${value} MOCK`);
+    expect(result).toMatchSnapshot();
+  });
+
+  test('transformValues async', async () => {
+    const result = await core.transformValues(complexJSON, async (value) => `${value} MOCK`);
+    expect(result).toMatchSnapshot();
   });
 });